// src/acct-creation/interface/sbacc.interface.ts
export interface SBAccount {
    cifId:          string
    schmCode:       string
    schmType:       string
    acctCrncy:      string
    branchId:       string
    glSubHeadCode:  string
    acctName:       string
    acctShortName:  string
}
  
  
export interface SBAccountRes {
    status: string
    message: string
}

  